Partial document support
Partially loaded documents will be tracked, and on save only those attributes
will be saved that have been loaded.

This is an early prototype. It works with the select() method of the criteria
and does not support EMongoCursor. Partial loading only works for the
attributes of the main document and not for embedded documents.